@@ -1,24 +1,91 @@
 <template>
-  <v-system-bar v-if="titleBar" class="title-bar" height="22" app status>
+  <v-system-bar
+    v-if="titleBar"
+    class="title-bar user-select-none"
+    :app="app"
+    :absolute="!app"
+    height="22"
+    @dblclick="handleDoubleClick"
+  >
     <v-spacer />
-    <span>{{ title }}</span>
+    <span class="caption text-truncate">Picty</span>
     <v-spacer />
   </v-system-bar>
 </template>
 
-<script>
-import { mapGetters, mapState } from 'vuex'
+<script lang="ts">
+import { remote } from 'electron'
+import {
+  defineComponent,
+  reactive,
+  computed,
+  onMounted,
+  onUnmounted,
+} from '@vue/composition-api'
 
-export default {
-  computed: {
-    ...mapState(['title']),
-    ...mapGetters(['titleBar'])
-  }
+type Props = {
+  app: boolean
 }
+
+export default defineComponent({
+  props: {
+    app: {
+      type: Boolean,
+      default: true,
+    },
+  },
+  setup(_props: Props) {
+    const state = reactive({
+      fullScreen: false,
+    })
+
+    const titleBar = computed(() => {
+      return process.platform === 'darwin' && !state.fullScreen
+    })
+
+    // @see https://github.com/electron/electron/issues/16385
+    const handleDoubleClick = () => {
+      const doubleClickAction = remote.systemPreferences.getUserDefault(
+        'AppleActionOnDoubleClick',
+        'string'
+      )
+      const win = remote.getCurrentWindow()
+      if (doubleClickAction === 'Minimize') {
+        win.minimize()
+      } else if (doubleClickAction === 'Maximize') {
+        if (win.isMaximized()) {
+          win.unmaximize()
+        } else {
+          win.maximize()
+        }
+      }
+    }
+    const handleFullScreenChange = () => {
+      state.fullScreen = !!document.fullscreenElement
+    }
+
+    onMounted(() => {
+      document.body.addEventListener('fullscreenchange', handleFullScreenChange)
+    })
+
+    onUnmounted(() => {
+      document.body.removeEventListener(
+        'fullscreenchange',
+        handleFullScreenChange
+      )
+    })
+
+    return {
+      titleBar,
+      handleDoubleClick,
+    }
+  },
+})
 </script>
 
-<style scoped lang="scss">
+<style lang="scss" scoped>
 .title-bar {
+  padding: 0 72px;
   -webkit-app-region: drag;
 }
 </style>
