@@ -1,102 +1,113 @@
 <template>
   <v-layout class="virtual-data-iterator" column>
-    <slot name="header" />
-    <v-container :class="classes" fluid pa-0 overflow-hidden>
-      <slot v-if="loading" name="progress" />
+    <v-container :class="classes" fluid pa-0 overflow-hidden fill-height>
       <v-data-iterator
         ref="iterator"
         v-model="model"
         v-bind="$attrs"
-        class="fill-height"
-        :pagination.sync="paginationModel"
+        class="fill-height flex-grow-1"
+        :options.sync="optionsModel"
         :items="renderItems"
-        :disable-initial-sort="true"
-        content-tag="v-layout"
+        disable-sort
+        disable-pagination
         row
         wrap
+        tabindex="-1"
       >
-        <template slot="item" slot-scope="props">
-          <v-flex
-            v-if="props.index === 0"
-            :style="{ height: `${padding.top}px` }"
-            class="pa-0"
-            xs12
-          />
-          <slot v-bind="props" name="items" />
-          <v-flex
-            v-if="props.index === renderItems.length - 1"
-            :style="{ height: `${padding.bottom}px` }"
-            class="pa-0"
-            xs12
-          />
+        <template v-slot:header="props">
+          <div class="header">
+            <slot v-bind="props" name="header" />
+          </div>
         </template>
-        <slot slot="no-data" name="no-data" />
-        <slot slot="no-results" name="no-results" />
+        <template v-slot:default="props">
+          <v-row class="ma-0">
+            <v-col
+              :style="{ height: `${padding.top}px` }"
+              class="pa-0"
+              xs="12"
+              style="min-width: 100%;"
+            />
+            <template v-for="item in props.items">
+              <slot v-bind="{ item }" name="item" />
+            </template>
+            <v-col
+              :style="{ height: `${padding.bottom}px` }"
+              class="pa-0"
+              xs="12"
+              style="min-width: 100%;"
+            />
+          </v-row>
+        </template>
+        <template v-slot:loading>
+          <v-progress-linear indeterminate height="2" />
+          <div class="ma-3 body-2 grey--text">Loading items...</div>
+        </template>
       </v-data-iterator>
     </v-container>
   </v-layout>
 </template>
 
 <script>
-import viewport from '~/utils/viewport'
+// TODO: typescriptize
+import * as viewport from '~/utils/viewport'
 
 export default {
   props: {
     value: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     pagination: {
       type: Object,
-      default: () => ({})
+      default: () => ({}),
     },
     items: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     estimatedHeight: {
       type: Number,
-      default: 48
+      default: 48,
     },
     threshold: {
       type: Number,
-      default: 0
-    },
-    loading: {
-      type: Boolean,
-      default: false
+      default: 0,
     },
     itemKey: {
       type: String,
-      default: 'id'
+      default: 'id',
     },
     sizes: {
       type: [Number, Array],
-      default: 6
+      default: 6,
     },
     containerClass: {
       type: [String],
-      default: ''
-    }
+      default: '',
+    },
+    stickyHeaders: {
+      type: Boolean,
+      default: false,
+    },
   },
   data() {
     return {
       scrolling: false,
       padding: {
         top: 0,
-        bottom: 0
+        bottom: 0,
       },
-      renderItems: []
+      renderItems: [],
     }
   },
   computed: {
-    paginationModel: {
+    optionsModel: {
       get() {
         return this.pagination
       },
       set(value) {
         this.$emit('update:pagination', value)
-      }
+      },
     },
     model: {
       get() {
@@ -104,25 +115,27 @@ export default {
       },
       set(value) {
         this.$emit('input', value)
-      }
+      },
     },
     classes() {
       return {
         [this.containerClass]: true,
-        scrolling: this.scrolling
+        'sticky-headers': this.stickyHeaders,
+        scrolling: this.scrolling,
       }
     },
     calculatedSizes() {
       return Array.isArray(this.sizes) ? this.sizes : Array(5).fill(this.sizes)
-    }
+    },
   },
   watch: {
     items() {
       this.adjustItems()
-    }
+    },
   },
   mounted() {
     this.container = this.$el.querySelector('.v-data-iterator')
+    this.container.classList.add('scrollbar')
     this.container.addEventListener('scroll', this.onScroll)
     this.observer = new ResizeObserver(this.onResize)
     this.observer.observe(this.container)
@@ -166,7 +179,7 @@ export default {
         top: firstIndex * this.estimatedHeight,
         bottom:
           (Math.ceil(this.items.length / size) - lastIndex) *
-          this.estimatedHeight
+          this.estimatedHeight,
       }
       this.renderItems = this.items.slice(firstIndex * size, lastIndex * size)
 
@@ -178,59 +191,22 @@ export default {
     onScroll(e) {
       this.adjustItems()
       this.$emit('scroll', e)
-    }
-  }
+    },
+  },
 }
 </script>
 
-<style scoped lang="scss">
+<style lang="scss" scoped>
 .virtual-data-iterator > .container {
   position: relative;
-  .v-progress-linear {
-    left: 0;
-    margin: 0;
-    position: absolute;
-    right: 0;
-    top: 0;
-  }
   .v-data-iterator {
     overflow-y: scroll;
-    &::-webkit-scrollbar {
-      width: 14px;
-    }
-    &::-webkit-scrollbar-thumb {
-      background-color: #eee;
-      &:hover {
-        background-color: #ddd;
-      }
-      &:active {
-        background-color: #ccc;
-      }
-    }
-    &:before {
-      box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
-      content: '';
-      left: 0;
-      position: absolute;
-      right: 14px;
-      top: -10px;
-      z-index: 1;
-    }
-    /deep/ .layout {
-      margin: 0px !important;
-    }
+    text-align: center;
   }
-  &.scrolling .v-data-iterator:before {
-    height: 10px;
-  }
-}
-.theme--dark .virtual-data-iterator .v-data-iterator::-webkit-scrollbar-thumb {
-  background-color: #424242 !important;
-  &:hover {
-    background-color: #505050 !important;
-  }
-  &:active {
-    background-color: #616161 !important;
+  ::v-deep .header {
+    position: sticky;
+    top: 0;
+    z-index: 1;
   }
 }
 </style>
