@@ -4,74 +4,73 @@
     v-model="model"
     v-bind="$attrs"
     class="virtual-data-table"
-    :pagination.sync="paginationModel"
+    :options.sync="optionsModel"
     :items="renderItems"
     :class="classes"
-    :disable-initial-sort="true"
+    disable-sort
+    disable-pagination
   >
-    <template slot="headers" slot-scope="props">
-      <slot v-bind="props" name="headers" />
+    <template v-slot:header="props">
+      <slot v-bind="props" name="header" />
     </template>
-    <template slot="items" slot-scope="props">
+    <template v-slot:item="props">
       <tr v-if="props.index === 0" :style="{ height: `${padding.top}px` }" />
-      <slot v-bind="props" name="items" />
+      <slot v-bind="props" name="item" />
       <tr
         v-if="props.index === renderItems.length - 1"
         :style="{ height: `${padding.bottom}px` }"
       />
     </template>
-    <slot slot="progress" name="progress" />
-    <slot slot="no-data" name="no-data" />
-    <slot slot="no-results" name="no-results" />
   </v-data-table>
 </template>
 
 <script>
+// TODO: typescriptize
 export default {
   props: {
     value: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     pagination: {
       type: Object,
-      default: () => ({})
+      default: () => ({}),
     },
     items: {
       type: Array,
-      default: () => []
+      default: () => [],
     },
     estimatedHeight: {
       type: Number,
-      default: 48
+      default: 48,
     },
     threshold: {
       type: Number,
-      default: 0
+      default: 0,
     },
     stickyHeaders: {
       type: Boolean,
-      default: false
-    }
+      default: false,
+    },
   },
   data() {
     return {
       scrolling: false,
       padding: {
         top: 0,
-        bottom: 0
+        bottom: 0,
       },
-      renderItems: []
+      renderItems: [],
     }
   },
   computed: {
-    paginationModel: {
+    optionsModel: {
       get() {
         return this.pagination
       },
       set(value) {
         this.$emit('update:pagination', value)
-      }
+      },
     },
     model: {
       get() {
@@ -79,22 +78,23 @@ export default {
       },
       set(value) {
         this.$emit('input', value)
-      }
+      },
     },
     classes() {
       return {
         'sticky-headers': this.stickyHeaders,
-        scrolling: this.scrolling
+        scrolling: this.scrolling,
       }
-    }
+    },
   },
   watch: {
     items() {
       this.adjustItems()
-    }
+    },
   },
   mounted() {
-    this.container = this.$el.querySelector('.v-table__overflow')
+    this.container = this.$el.querySelector('.v-data-table__wrapper')
+    this.container.classList.add('scrollbar')
     this.container.addEventListener('scroll', this.onScroll)
     this.observer = new ResizeObserver(this.onResize)
     this.observer.observe(this.container)
@@ -134,7 +134,7 @@ export default {
       this.scrolling = scrollTop > 0
       this.padding = {
         top: firstIndex * this.estimatedHeight,
-        bottom: (this.items.length - lastIndex) * this.estimatedHeight
+        bottom: (this.items.length - lastIndex) * this.estimatedHeight,
       }
       this.renderItems = this.items.slice(firstIndex, lastIndex)
 
@@ -146,30 +146,20 @@ export default {
     onScroll(e) {
       this.adjustItems()
       this.$emit('scroll', e)
-    }
-  }
+    },
+  },
 }
 </script>
 
-<style scoped lang="scss">
+<style lang="scss" scoped>
 .virtual-data-table.sticky-headers {
-  /deep/ .v-table__overflow {
+  ::v-deep .v-data-table__wrapper {
     height: 100%;
     overflow-y: scroll;
-    &::-webkit-scrollbar {
-      width: 14px;
-    }
-    &::-webkit-scrollbar-thumb {
-      background-color: #eee;
-      &:hover {
-        background-color: #ddd;
-      }
-      &:active {
-        background-color: #ccc;
-      }
-    }
-    .v-datatable {
+    background: inherit;
+    table {
       table-layout: fixed;
+      background: inherit;
       > thead {
         background: inherit;
         > tr {
@@ -180,39 +170,9 @@ export default {
             top: 0;
             z-index: 1;
           }
-          &.v-datatable__progress > th {
-            top: 56px;
-            z-index: 0;
-            &:after {
-              bottom: 0;
-              box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
-              content: '';
-              left: 0;
-              position: absolute;
-              width: 100%;
-            }
-          }
         }
       }
     }
-  }
-  &.scrolling /deep/ .v-datatable > thead > tr {
-    border-bottom: none;
-    &.v-datatable__progress > th:after {
-      height: 10px;
-    }
-  }
-}
-.theme--dark
-  .virtual-data-table.sticky-headers
-  /deep/
-  .v-table__overflow::-webkit-scrollbar-thumb {
-  background-color: #424242 !important;
-  &:hover {
-    background-color: #505050 !important;
-  }
-  &:active {
-    background-color: #616161 !important;
   }
 }
 </style>
